create magement service
