+++ app/session/index.ts
@@ -1,5 +1,5 @@
 import type { SessionState } from "../types";
-import { zombieSessionMonitor } from "./health";
+import { zombieSessionMonitor } from "./service";
 
 // === Configuracao de reconnect ===
 export const HEALTH_CHECK_INTERVAL_MS = 60000; // 1 minuto
@@ -24,7 +24,4 @@ export const sessions: Record<string, SessionState> = {};
 
 setInterval(zombieSessionMonitor, HEALTH_CHECK_INTERVAL_MS);
 
-export * from "./status";
-export * from "./health";
-export * from "./handlers";
-export * from "./management";
+export * from "./service";
